Bug in fixQuality fixed. Status bar up top in QtAOG now shows the fix quality and the age.

+++ formgps_position.cpp
@@ -1159,7 +1159,7 @@ void FormGPS::UpdateFixPosition()
     aog->setProperty("altitude", pn.altitude);
     aog->setProperty("hdop", pn.hdop);
     aog->setProperty("age", pn.age);
-    aog->setProperty("fixQuality", pn.fixQuality);
+    aog->setProperty("fixQuality", (int)pn.fixQuality);
     aog->setProperty("satellitesTracked", pn.satellitesTracked);
     aog->setProperty("imuHeading", ahrs.imuHeading);
     aog->setProperty("angVel", ahrs.angVel);
